Rename dnd dice selector to numbered dice selector
