--- src/components/dnd-dice-selector/dnd-dice-selector.scss
@@ -1,32 +1,0 @@
-/**
- * @author Michael Gamlem III
- * @copyright This file is subject to the terms and conditions defined in file 'LICENSE', which is part of the source code for this project.
- * @format
- */
-
-.dnd-dice-selector {
-	width: 40%;
-	min-height: 370px;
-	display: flex;
-	flex-direction: column;
-	justify-content: space-between;
-
-	.dice-options-container {
-		display: flex;
-
-		.dice-option-column {
-			padding: 0 4px;
-			max-width: 150px;
-
-			.dice-option {
-				margin-bottom: 10px;
-			}
-		}
-	}
-
-	.button-container {
-		button {
-			margin-right: 4px;
-		}
-	}
-}
